requirements.txt
asgiref==3.11.1
django==4.2.28
djangorestframework==3.16.1
pillow==11.3.0
pymysql==1.1.2
python-decouple==3.8
qrcode==8.2
sqlparse==0.5.5
typing-extensions==4.15.0
